C++ College Programming
StudentWorks is a very simple C++ program I've written which serves as a
hub connected to every other C++ program I wrote while I was in college.
StudentWorks is an early example of my basic knowledge of the C++
language and Microsoft Visual Studio. It is occasionally being updated
for greater usability.